Conflicts:
	titan-core/src/main/java/com/thinkaurelius/titan/diskstorage/keycolumnvalue/cache/KCVSCache.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/database/EdgeSerializer.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/database/IndexSerializer.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/database/StandardTitanGraph.java
	titan-core/src/main/java/com/thinkaurelius/titan/graphdb/log/StandardTransactionLogProcessor.java
	titan-test/src/main/java/com/thinkaurelius/titan/graphdb/TitanGraphTest.java
	titan-test/src/main/java/com/thinkaurelius/titan/graphdb/TitanIndexTest.java

+++ titan-cassandra/src/main/java/com/thinkaurelius/titan/diskstorage/cassandra/astyanax/AstyanaxKeyColumnValueStore.java
@@ -156,7 +156,7 @@ public class AstyanaxKeyColumnValueStore implements KeyColumnValueStore {
                 case TIMESTAMP:
                     return element.getTimestamp();
                 case TTL:
-                    return (long) element.getTtl();
+                    return element.getTtl();
                 default:
                     throw new UnsupportedOperationException("Unsupported meta data: " + meta);
             }
